requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
cffi==1.17.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==43.0.0
daphne==4.1.2
distlib==0.3.8
distro==1.9.0
Django==5.0.7
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-timezone-field==7.0
djangorestframework==3.15.2
dnspython==2.6.1
docx2pdf==0.1.8
eventlet==0.36.1
filelock==3.15.4
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
jiter==0.5.0
kombu==5.4.0
lxml==5.2.2
msgpack==1.0.8
numpy==2.1.0
openai==1.40.2
packaging==24.1
pandas==2.2.2
pandoc==2.3
pdfkit==1.0.0
pillow==10.4.0
platformdirs==4.2.2
plumbum==1.8.3
ply==3.11
prompt_toolkit==3.0.47
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pyOpenSSL==24.2.1
pypandoc==1.13
PyPDF2==3.0.1
pytesseract==0.3.10
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
pytz==2024.1
redis==5.0.8
requests==2.32.3
service-identity==24.1.0
setuptools==74.0.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
tqdm==4.66.5
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
virtualenv==20.26.3
wcwidth==0.2.13
wkhtmltopdf==0.2
zope.interface==7.0.3
